(WIP) Update to Groovy 3.0.4 for testing on newer Java versions
Since GMavenPlus 1.9.0 Java 14 and Java 15 is supported. This version of
GMavenPlus requires Groovy 3.x, so also update Groovy to latest version.
Java 7 is not supported any more.

For other JVM languages I propose that we focus on the last stable
release, so Groovy 3.x should be ok as a test target.